Move these tests over
