+++ Soundboard.Tests/Mp3SoundTests.cs
@@ -29,31 +29,5 @@ namespace Soundboard.Tests
 
             Subject.LoadFile(theSoundFile);
         }
-
-        [Test]
-        [TestCase(@"example-sounds\mpthreetest.mp3")]
-        public void It_can_block_until_playback_is_complete(string pathToSoundFile)
-        {
-            var theSoundFile = GetTheProjectRelativePathForThisProjectFile(pathToSoundFile);
-
-            Subject.LoadFile(theSoundFile);
-            Subject.PlayToCompletion();
-
-            Subject.CurrentPlaybackState.ShouldEqual(PlaybackState.Stopped);
-        }
-
-        [Test]
-        [TestCase(@"example-sounds\mpthreetest.mp3")]
-        public void It_can_stop_playback_before_playback_is_complete(string pathToSoundFile)
-        {
-            var theSoundFile = GetTheProjectRelativePathForThisProjectFile(pathToSoundFile);
-
-            Subject.LoadFile(theSoundFile);
-            Subject.Play();
-            Thread.Sleep(500);
-            Subject.Stop();
-
-            Subject.CurrentPlaybackState.ShouldEqual(PlaybackState.Stopped);
-        }
     }
 }
