+++ Soundboard.Tests/PlayerTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Threading;
 using AutoMoq.Helpers;
 using NAudio.Wave;
 using NUnit.Framework;
@@ -25,6 +26,32 @@ namespace Soundboard.Tests
             Subject.Play();
 
             Subject.CurrentPlaybackState.ShouldEqual(PlaybackState.Playing);
+        }
+
+        [Test]
+        [TestCase(@"example-sounds\mpthreetest.mp3")]
+        public void It_can_block_until_playback_is_complete(string pathToSoundFile)
+        {
+            var theSoundFile = GetTheProjectRelativePathForThisProjectFile(pathToSoundFile);
+
+            Subject.LoadFile(theSoundFile);
+            Subject.PlayToCompletion();
+
+            Subject.CurrentPlaybackState.ShouldEqual(PlaybackState.Stopped);
+        }
+
+        [Test]
+        [TestCase(@"example-sounds\mpthreetest.mp3")]
+        public void It_can_stop_playback_before_playback_is_complete(string pathToSoundFile)
+        {
+            var theSoundFile = GetTheProjectRelativePathForThisProjectFile(pathToSoundFile);
+
+            Subject.LoadFile(theSoundFile);
+            Subject.Play();
+            Thread.Sleep(500);
+            Subject.Stop();
+
+            Subject.CurrentPlaybackState.ShouldEqual(PlaybackState.Stopped);
         }
     }
 }
